// apps/console/src/pages/organizations/obligations/__generated__/ObligationsPageFragment.graphql.ts
export type ObligationStatus = "CLOSED" | "IN_PROGRESS" | "OPEN";
import { FragmentRefs } from "relay-runtime";
export type ObligationsPageFragment$data = {
  readonly id: string;
  readonly obligations: {
    readonly __id: string;
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly actionsToBeImplemented: string | null | undefined;
        readonly area: string | null | undefined;
        readonly createdAt: any;
        readonly dueDate: any | null | undefined;
        readonly id: string;
        readonly lastReviewDate: any | null | undefined;
        readonly owner: {
          readonly fullName: string;
          readonly id: string;
        };
        readonly referenceId: string;
        readonly regulator: string | null | undefined;
        readonly requirement: string | null | undefined;
        readonly snapshotId: string | null | undefined;
        readonly source: string | null | undefined;
        readonly sourceId: string | null | undefined;
        readonly status: ObligationStatus;
        readonly updatedAt: any;
      };
    }>;
    readonly pageInfo: {
      readonly endCursor: any | null | undefined;
      readonly hasNextPage: boolean;
    };
    readonly totalCount: number;
  };
  readonly " $fragmentType": "ObligationsPageFragment";
};
export type ObligationsPageFragment$key = {
  readonly " $data"?: ObligationsPageFragment$data;
  readonly " $fragmentSpreads": FragmentRefs<"ObligationsPageFragment">;
};